Add null-check and string/int value validation to parameters, add test case for validation
