+++ test/README.md
@@ -1,0 +1,3 @@
+# Test Setup
+This module uses Terratest and Go to test the infrastructure.
+It uses Localstack to mock AWS services.
