+++ variables.tf
@@ -11,6 +11,11 @@ variable "name" {
 variable "parameters" {
   type        = map(string)
   description = "Map of SSM ParameterStore parameters to store - by default, /$var.env/$var.name/*"
+
+  validation {
+    condition     = alltrue([for v in var.parameters : (v != null && v != "" && can(v) || can(int(v)))])
+    error_message = "All values in the 'parameters' map must be non-null, non-empty strings."
+  }
 }
 
 locals {
